requirements.txt
pandas
requests
python-dotenv
prefect
duckdb
pyarrow